Add DJI Ronin focus motor support
While I was able to control focus motor speed directly when doing
preliminary testing with the RS3, it seems like you're only able to send a
target position for the focus motor on the RSC 2. The result a much more
complicated control system that doesn't give perfectly smooth results, but
it'll have to do.

I'll potentially need to differentiate between different Ronin versions in
the future, but I'll look into that the next time I get access to a RS3.