Read Lumix event stream
It has occurred to me that the reason my Lumix connections keep becoming
unresponsive after a couple of hours is probably because I'm not actually
reading from the event socket and a buffer is filling up and eventually
blocking new packets. At least, that's my hunch.

Also added a timeout when reading the command stream. If I still run into
issues, this might help diagnose what's going on.